Added profile configuration which produces profile statistics that can in turn be analysed by gprof. To do that, run your program first alone, then rerun it with gprof, e.g.   your_program   gprof your_program
Currently, most time is spend in the inflate algorithm, which could be sped up by a cache, similar to what git uses. We might make it configurable, as git uses a static cache with X fixed entries. Depending on the task, it might be useful to allow custom sizes with memory limits, to increase the chance of cache hits. The implementation must be done carefully as building up the cache could as well cost a lot extra and might not even make up for the costs if we have many cache misses and thus add/remove entries without ever using them.
The test also shows that apply-delta only costs a fraction of the time, so this is not where any time can be saved.
Also , it might be worth adding an alternate implementation which prevents two big buffers + one delta to be in memory at once by merging deltas before applying them (see gitdb)